fix rename service
