+++ function-download-image/src/service/download-image.service.ts
